+++ dmutils/logging.py
@@ -1,12 +1,10 @@
 from __future__ import absolute_import
 import logging
-import uuid
 import sys
 import re
 from itertools import product
 
 from flask import request, current_app
-from flask.wrappers import Request
 from flask.ctx import has_request_context
 
 from pythonjsonlogger.jsonlogger import JsonFormatter as BaseJSONFormatter
@@ -22,16 +20,9 @@ def init_app(app):
     app.config.setdefault('DM_LOG_LEVEL', 'INFO')
     app.config.setdefault('DM_APP_NAME', 'none')
     app.config.setdefault('DM_LOG_PATH', './log/application.log')
-    app.config.setdefault('DM_REQUEST_ID_HEADER', 'DM-Request-ID')
-    app.config.setdefault('DM_DOWNSTREAM_REQUEST_ID_HEADER', '')
-
-    app.request_class = CustomRequest
 
     @app.after_request
     def after_request(response):
-        request_id_header = current_app.config['DM_REQUEST_ID_HEADER']
-        response.headers[request_id_header] = request.request_id
-
         current_app.logger.info('%s %s %s',
                                 request.method,
                                 request.url,
@@ -50,26 +41,6 @@ def init_app(app):
         logger.setLevel(loglevel)
 
     app.logger.info("Logging configured")
-
-
-class CustomRequest(Request):
-    _request_id = None
-
-    @property
-    def request_id(self):
-        if self._request_id is None:
-            self._request_id = self._get_request_id(
-                current_app.config['DM_REQUEST_ID_HEADER'],
-                current_app.config['DM_DOWNSTREAM_REQUEST_ID_HEADER'])
-        return self._request_id
-
-    def _get_request_id(self, request_id_header, downstream_header):
-        if request_id_header in self.headers:
-            return self.headers.get(request_id_header)
-        elif downstream_header and downstream_header in self.headers:
-            return self.headers.get(downstream_header)
-        else:
-            return str(uuid.uuid4())
 
 
 def configure_handler(handler, app, formatter):
