+++ tests/test_apiclient.py
@@ -13,6 +13,7 @@ from dmutils.apiclient import APIError, HTTPError, InvalidResponse
 from dmutils.apiclient.errors import REQUEST_ERROR_STATUS_CODE
 from dmutils.apiclient.errors import REQUEST_ERROR_MESSAGE
 from dmutils.audit import AuditTypes
+from dmutils import request_id
 
 
 @pytest.yield_fixture
@@ -278,6 +279,7 @@ class TestDataApiClient(object):
     def test_request_id_is_added_if_available(
             self, data_client, rmock, app_with_logging):
         headers = {'DM-Request-Id': 'generated'}
+        request_id.init_app(app_with_logging)
         with app_with_logging.test_request_context('/', headers=headers):
             rmock.get(
                 "http://baseurl/_status",
